add a deep_merge_cache mixin
gets the inheritence chain correct
