+++ lib/chef/node/attribute.rb
@@ -17,6 +17,7 @@
 # limitations under the License.
 #
 
+require "chef/node/mixin/deep_merge_cache"
 require "chef/node/mixin/immutablize_hash"
 require "chef/node/mixin/state_tracking"
 require "chef/node/immutable_collections"
@@ -36,8 +37,17 @@ class Chef
     class Attribute < Mash
 
       include Immutablize
-
+      # FIXME:  what is include Enumerable doing up here, when down below we delegate
+      # most of the Enumerable/Hash things to the underlying merged ImmutableHash.  That
+      # is, in fact, the correct, thing to do, while including Enumerable to try to create
+      # a hash-like API gets lots of things wrong because of the difference between the
+      # Hash `each do |key, value|` vs the Array-like `each do |value|` API that Enumerable
+      # expects.  This include should probably be deleted?
       include Enumerable
+
+      include Chef::Node::Mixin::DeepMergeCache
+      include Chef::Node::Mixin::StateTracking
+      include Chef::Node::Mixin::ImmutablizeHash
 
       # List of the component attribute hashes, in order of precedence, low to
       # high.
@@ -177,19 +187,6 @@ class Chef
        # return the automatic level attribute component
       attr_reader :automatic
 
-       # This is used to track the top level key as we descend through method chaining into
-       # a precedence level (e.g. node.default['foo']['bar']['baz']= results in 'foo' here).  We
-       # need this so that when we hit the end of a method chain which results in a mutator method
-       # that we can invalidate the whole top-level deep merge cache for the top-level key.  It is
-       # the responsibility of the accessor on the Chef::Node object to reset this to nil, and then
-       # the first VividMash#[] call can ||= and set this to the first key we encounter.
-      attr_accessor :top_level_breadcrumb
-
-       # Cache of deep merged values by top-level key.  This is a simple hash which has keys that are the
-       # top-level keys of the node object, and we save the computed deep-merge for that key here.  There is
-       # no cache of subtrees.
-      attr_accessor :deep_merge_cache
-
       def initialize(normal, default, override, automatic)
         @default = VividMash.new(self, default)
         @env_default = VividMash.new(self, {})
@@ -204,12 +201,7 @@ class Chef
         @force_override = VividMash.new(self, {})
 
         @automatic = VividMash.new(self, automatic)
-
-        @merged_attributes = nil
-        @combined_override = nil
-        @combined_default = nil
-        @top_level_breadcrumb = nil
-        @deep_merge_cache = {}
+        super()
       end
 
        # Debug what's going on with an attribute. +args+ is a path spec to the
@@ -236,20 +228,6 @@ class Chef
           [component.to_s.sub(/^@/, ""), value]
         end
       end
-
-       # Invalidate a key in the deep_merge_cache.  If called with nil, or no arg, this will invalidate
-       # the entire deep_merge cache.  In the case of the user doing node.default['foo']['bar']['baz']=
-       # that eventually results in a call to reset_cache('foo') here.  A node.default=hash_thing call
-       # must invalidate the entire cache and re-deep-merge the entire node object.
-      def reset_cache(path = nil)
-        if path.nil?
-          @deep_merge_cache = {}
-        else
-          deep_merge_cache.delete(path.to_s)
-        end
-      end
-
-      alias :reset :reset_cache
 
        # Set the cookbook level default attribute component to +new_data+.
       def default=(new_data)
@@ -415,36 +393,6 @@ class Chef
         write(:force_override, *args, value)
       end
 
-      # method-style access to attributes
-
-      def read(*path)
-        merged_attributes.read(*path)
-      end
-
-      def read!(*path)
-        merged_attributes.read!(*path)
-      end
-
-      def exist?(*path)
-        merged_attributes.exist?(*path)
-      end
-
-      def write(level, *args, &block)
-        self.send(level).write(*args, &block)
-      end
-
-      def write!(level, *args, &block)
-        self.send(level).write!(*args, &block)
-      end
-
-      def unlink(level, *path)
-        self.send(level).unlink(*path)
-      end
-
-      def unlink!(level, *path)
-        self.send(level).unlink!(*path)
-      end
-
        #
        # Accessing merged attributes.
        #
@@ -486,24 +434,39 @@ class Chef
         write(:normal, *args) if read(*args[0...-1]).nil?
       end
 
-      def [](key)
-        if deep_merge_cache.has_key?(key.to_s)
-          # return the cache of the deep merged values by top-level key
-          deep_merge_cache[key.to_s]
-        else
-          # save all the work of computing node[key]
-          deep_merge_cache[key.to_s] = merged_attributes(key)
-        end
-      end
-
-      def []=(key, value)
-        raise Exceptions::ImmutableAttributeModification
-      end
-
       def has_key?(key)
         COMPONENTS.any? do |component_ivar|
           instance_variable_get(component_ivar).has_key?(key)
         end
+      end
+       # method-style access to attributes (has to come after the prepended ImmutablizeHash)
+
+      def read(*path)
+        merged_attributes.read(*path)
+      end
+
+      def read!(*path)
+        merged_attributes.read!(*path)
+      end
+
+      def exist?(*path)
+        merged_attributes.exist?(*path)
+      end
+
+      def write(level, *args, &block)
+        self.send(level).write(*args, &block)
+      end
+
+      def write!(level, *args, &block)
+        self.send(level).write!(*args, &block)
+      end
+
+      def unlink(level, *path)
+        self.send(level).unlink(*path)
+      end
+
+      def unlink!(level, *path)
+        self.send(level).unlink!(*path)
       end
 
       alias :attribute? :has_key?
@@ -632,14 +595,13 @@ class Chef
           Chef::Mixin::DeepMerge.deep_merge(component_value, merged)
         end
       end
+
+       # needed for __path
+      def convert_key(key)
+        key.kind_of?(Symbol) ? key.to_s : key
+      end
+
     end
 
-    # needed for __path
-    def convert_key(key)
-      key.kind_of?(Symbol) ? key.to_s : key
-    end
-
-    prepend Chef::Node::Mixin::StateTracking
-    prepend Chef::Node::Mixin::ImmutablizeHash
   end
 end
